+++ melos_onyx_workspace.iml
@@ -238,6 +238,9 @@
       <excludeFolder url="file://$MODULE_DIR$/apps/.flutter/packages/integration_test/integration_test_macos/.dart_tool" />
       <excludeFolder url="file://$MODULE_DIR$/apps/.flutter/packages/integration_test/integration_test_macos/.pub" />
       <excludeFolder url="file://$MODULE_DIR$/apps/.flutter/packages/integration_test/integration_test_macos/build" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/polytechcolloscopeclient/.dart_tool" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/polytechcolloscopeclient/.pub" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/polytechcolloscopeclient/build" />
     </content>
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" name="Dart SDK" level="project" />
