+++ packages/polytechcolloscopeclient/lib/polytechcolloscopeclient.dart
@@ -1,1 +1,2 @@
+export 'src/consts.dart';
 export 'src/polytechcolloscopeclient_base.dart';
