+++ packages/polytechcolloscopeclient/lib/src/polytechcolloscopeclient_base.dart
@@ -1,51 +1,17 @@
 import 'package:beautiful_soup_dart/beautiful_soup.dart';
-import 'package:equatable/equatable.dart';
+import 'package:html/dom.dart';
 import 'package:requests_plus/requests_plus.dart';
 
-import 'package:html/dom.dart';
-
-enum Year {
-  first,
-  second,
-}
+import 'colloscope_datastructs.dart';
+import 'consts.dart';
 
 class PolytechColloscopeClient {
-  static const String userAgent =
-      "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:96.0) Gecko/20100101 Firefox/96.0";
-
-  // HTTPS parce que le https est pas géré
-  static const String _base = "http://math.univ-lyon1.fr/colles/";
-
-  static const kholleURL = {
-    Year.first: "$_base?page=colles_1A",
-    Year.second: "$_base?page=colles_2A"
-  };
-
-  static const khollesStudentURL = {
-    Year.first: "$_base?page=colles_1A&id_etudiant=:id",
-    Year.second: "$_base?page=colles_2A&id_etudiant=:id"
-  };
-
-  static const Map monthsTranslation = {
-    "janvier": DateTime.january,
-    "février": DateTime.february,
-    "mars": DateTime.march,
-    "avril": DateTime.april,
-    "mai": DateTime.may,
-    "juin": DateTime.june,
-    "juillet": DateTime.july,
-    "août": DateTime.august,
-    "septembre": DateTime.september,
-    "octobre": DateTime.october,
-    "novembre": DateTime.november,
-    "décembre": DateTime.december
-  };
-
   final String username, password;
+
   PolytechColloscopeClient(this.username, this.password);
 
   Future<List<Student>> fetchStudents(Year year) async {
-    var page = await RequestsPlus.get(kholleURL[year]!,
+    var page = await RequestsPlus.get(Consts.kholleURL[year]!,
         userName: username, password: password);
 
     BeautifulSoup bs = BeautifulSoup(page.body);
@@ -71,7 +37,8 @@ class PolytechColloscopeClient {
     // TODO : Add a check to verify the studentid (look 5 lines below)
 
     var page = await RequestsPlus.get(
-        khollesStudentURL[year]!.replaceFirst(":id", studentId.toString()),
+        Consts.khollesStudentURL[year]!
+            .replaceFirst(":id", studentId.toString()),
         userName: username,
         password: password);
 
@@ -102,31 +69,33 @@ class PolytechColloscopeClient {
     var secondTd = e.children.elementAtOrNull(1);
 
     var hourAndMinute =
-    secondTd?.children.first.innerHtml.replaceFirst("&nbsp;", "").trim();
+        secondTd?.children.first.innerHtml.replaceFirst("&nbsp;", "").trim();
 
     var secondDiv = secondTd?.children.elementAtOrNull(1);
     var kholleur = secondDiv?.find("a")?.text.trim();
 
-    // This hold the subject of the kholle (and maybe the room because this website html is probably the worst I've ever seen)
+    // This holds the subject of the kholle (and maybe the room because this website html is probably the worst I've ever seen)
     var text = "";
 
-    var divText = secondDiv?.nodes.where((element) => element.runtimeType == Text);
+    var divText =
+        secondDiv?.nodes.where((element) => element.runtimeType == Text);
 
     var subject = divText?.first.text?.replaceAll(RegExp(r'[()]'), "").trim();
 
     String message;
     if (divText?.length == 2) {
-      message = divText?.last.text?.replaceAll(RegExp(r'[()]'), "").trim() ?? "";
+      message =
+          divText?.last.text?.replaceAll(RegExp(r'[()]'), "").trim() ?? "";
     } else {
       message = "";
     }
 
     var dateParsed = RegExp(r"(\d{1,2}) (.{3,9})").firstMatch(date);
     var day = dateParsed?.group(1);
-    var month = monthsTranslation[dateParsed?.group(2)];
+    var month = Consts.monthsTranslation[dateParsed?.group(2)];
 
     var hourAndMinutesParsed =
-    RegExp(r"(\d{1,2}) h (\d{2})").firstMatch(hourAndMinute!);
+        RegExp(r"(\d{1,2}) h (\d{2})").firstMatch(hourAndMinute!);
     var hour = hourAndMinutesParsed?.group(1);
     var minutes = hourAndMinutesParsed?.group(2);
 
@@ -135,56 +104,4 @@ class PolytechColloscopeClient {
 
     return Kholle(dateTime, subject!, message, kholleur);
   }
-}
-
-class Student extends Equatable {
-  final String name;
-  final String id;
-
-  Student(this.name, this.id);
-
-  @override
-  List<Object?> get props => [name, id];
-
-  @override
-  bool? get stringify => true;
-}
-
-class StudentColloscope extends Equatable {
-  final int studentId;
-  final int trinomeId;
-  final List kholles;
-
-  StudentColloscope(this.studentId, this.trinomeId, this.kholles);
-
-  @override
-  List<Object?> get props => [studentId, trinomeId, kholles];
-
-  @override
-  bool? get stringify => true;
-}
-
-class Kholle extends Equatable {
-  final DateTime date;
-  final String subject;
-  final String? message;
-  final String? kholleur;
-
-  Kholle(this.date, this.subject, this.message, this.kholleur);
-
-  @override
-  List<Object?> get props => [date, subject, message, kholleur];
-
-  @override
-  bool? get stringify => true;
-}
-
-Future<void> main() async {
-  PolytechColloscopeClient client = PolytechColloscopeClient("", "");
-
-  var students = await client.fetchStudents(Year.first);
-  print(students);
-
-  var colloscope = await client.getColloscope(Year.first, 828);
-  print(colloscope);
 }
