+++ packages/polytechcolloscopeclient/pubspec.yaml
@@ -9,6 +9,7 @@ environment:
 # Add regular dependencies here.
 dependencies:
   beautiful_soup_dart: ^0.3.0
+  dotenv: ^4.2.0
   equatable: ^2.0.5
   requests_plus: ^4.8.4
   # path: ^1.8.0
