+++ packages/polytechcolloscopeclient/test/polytechcolloscopeclient_test.dart
@@ -1,6 +1,23 @@
+import 'package:dotenv/dotenv.dart';
 import 'package:polytechcolloscopeclient/polytechcolloscopeclient.dart';
 import 'package:test/test.dart';
 
-void main() {
-  
+Future<void> main() async {
+  DotEnv env = DotEnv(includePlatformEnvironment: true);
+
+  env.load();
+  final String username = env['USERNAME'] ?? "";
+  final String password = env['PASSWORD'] ?? "";
+
+  if (username.isEmpty || password.isEmpty) {
+    fail("username or password were empty. check your envt variables");
+  }
+
+  PolytechColloscopeClient client = PolytechColloscopeClient("", "");
+
+  var students = await client.fetchStudents(Year.first);
+  print(students);
+
+  var colloscope = await client.getColloscope(Year.first, 828);
+  print(colloscope);
 }
